django-swg/requirements.txt
Django~=4.2.3
drf-spectacular~=0.26.1
djangorestframework~=3.14.0